+++ src/main/java/com/example/shinsekai/purchase/vo/in/OrderRequestVo.java
@@ -5,7 +5,6 @@ import com.example.shinsekai.purchase.entity.PurchaseStatus;
 import lombok.Getter;
 import lombok.ToString;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Getter
@@ -17,8 +16,8 @@ public class OrderRequestVo {
     private String address;
     private String giftCertificationUuid;
     private String couponUuid;
-    private int shipmentFee;
-    private double productTotalPrice;
+    private Double shipmentFee;
+    private Double productTotalPrice;
 
     private List<PurchaseProductRequestVo> orderProductList;
